MED-144 fix: separating patient notes from book patient appointment
