+++ apps/medon-fe/src/components/PatientCardCalendar/index.tsx
@@ -1,33 +1,88 @@
 import { useState } from 'react';
-import dayjs from 'dayjs';
 import { useTranslation } from 'react-i18next';
 import { Views, dayjsLocalizer, Event } from 'react-big-calendar';
+import AnimateHeight, { Height } from 'react-animate-height';
+import { SubmitHandler, useForm } from 'react-hook-form';
+import { useParams, useSearchParams } from 'react-router-dom';
+import { toast } from 'react-toastify';
+import { Input } from 'antd';
+import dayjs from 'dayjs';
+import { yupResolver } from '@hookform/resolvers/yup';
 import { useTheme } from 'styled-components';
 
+import { AddNote, Close } from 'assets/svgs/patientCard';
+import Plus from 'assets/svgs/plus_listcard.svg';
+
 import Link from 'components/Link';
+import { PatientNotes } from 'components/PatientNotes';
+import Button from 'components/Button';
+import { TextareaAntD } from 'components/common';
 
 import { eventsCard } from 'utils/mock/patientCalendar';
 import { getDateAndHourEvent } from 'utils/functions/getDateAndHourEvent';
 import { getDayPropGetter } from 'utils/functions/getDayPropGetter';
 import { getEventPropGetter } from 'utils/functions/getEventPropGetter';
-import { roles, routes } from 'utils/constants';
+import {
+  defaultOrder,
+  defaultPage,
+  defaultPageSize,
+  roles,
+  routes,
+} from 'utils/constants';
+import { toastConfig } from 'utils/toastConfig';
+import { options } from 'utils/constants/options/patientCardSelect';
+
+import { useDebounce } from 'hooks/useDebounce';
 
 import { useAppSelector } from 'redux/hooks';
+import {
+  useCreatePatientNoteMutation,
+  useGetPatientNotesQuery,
+} from 'redux/api/patientApi';
 
-import Plus from 'assets/svgs/plus_listcard.svg';
+import { addPatientNoteSchema } from 'validation/addPatientNoteSchema';
 
-import { StyledCalendar, StyledModal, Title } from './styles';
+import {
+  AddNoteForm,
+  Buttons,
+  StyledCalendar,
+  StyledModal,
+  StyledSelect,
+  Title,
+  Wrapper,
+} from './styles';
 import { useModal } from './hooks';
+import { SubmitAddNote } from './types';
 import 'react-big-calendar/lib/css/react-big-calendar.css';
 
 export function PatientCardCalendar() {
   const [event, setEvent] = useState<Event>();
+  const [height, setHeight] = useState<Height>(0);
+  const [textValue, setTextValue] = useState<string>('');
 
-  const theme = useTheme();
-  const { t } = useTranslation();
+  const { id } = useParams();
   const { role } = useAppSelector((state) => state.userState.user);
+  const { t } = useTranslation();
+  const [searchParams, setSearchParams] = useSearchParams();
+  const page = Number(searchParams.get('page')) || defaultPage;
+  const limit = Number(searchParams.get('limit')) || defaultPageSize;
+  const order = searchParams.get('order') || defaultOrder;
+  const theme = useTheme();
 
+  const text = useDebounce(textValue, 1000);
   const { hideModal, isVisible, showModal } = useModal(false);
+  const { data: notes, isFetching } = useGetPatientNotesQuery({
+    id,
+    order,
+    text,
+    page,
+    limit,
+  });
+  const [sendData, { isLoading: isNoteSending }] =
+    useCreatePatientNoteMutation();
+  const { handleSubmit, control, reset } = useForm<SubmitAddNote>({
+    resolver: yupResolver(addPatientNoteSchema),
+  });
 
   const localizer = dayjsLocalizer(dayjs);
   const dayPropGetter = getDayPropGetter(theme);
@@ -37,6 +92,18 @@ export function PatientCardCalendar() {
     showModal();
     setEvent(eventValue);
   }
+
+  const handleAddNote: SubmitHandler<SubmitAddNote> = ({ note }) => {
+    sendData({ note, patientId: id })
+      .unwrap()
+      .then(() => {
+        toast.success(t('patient-card.toast.success'), toastConfig);
+        reset();
+      })
+      .catch((err) => {
+        toast.error(err.data.message, toastConfig);
+      });
+  };
 
   return (
     <>
@@ -64,7 +131,6 @@ export function PatientCardCalendar() {
         onSelectEvent={handleEventSelect}
         popup
         selectable
-        timeslots={1}
         step={60}
       />
       <StyledModal
@@ -96,6 +162,67 @@ export function PatientCardCalendar() {
           </Link>
         </p>
       </StyledModal>
+      <Button
+        isfullwidth="true"
+        textcolor={theme.colors.blue_500}
+        bgcolor={theme.colors.blue_100}
+        onClick={() => setHeight(height === 0 ? 'auto' : 0)}
+      >
+        {t('patient-card.button')}
+      </Button>
+      <AnimateHeight height={height}>
+        <AddNoteForm onSubmit={handleSubmit(handleAddNote)}>
+          <TextareaAntD
+            name="note"
+            control={control}
+            minRows={6}
+            placeholder={`${t('patient-card.notes.placeholder')}`}
+          />
+          <Buttons>
+            <Button
+              textcolor={theme.colors.white}
+              bgcolor={theme.colors.btnGradient}
+              isLoading={isNoteSending}
+            >
+              {t('patient-card.notes.add-button')}
+              <AddNote />
+            </Button>
+            <Button
+              type="button"
+              textcolor={theme.colors.gray_700}
+              bgcolor={theme.colors.gray_400}
+              onClick={() => setHeight(height === 0 ? 'auto' : 0)}
+            >
+              <Close />
+              {t('patient-card.notes.cancel-button')}
+            </Button>
+          </Buttons>
+        </AddNoteForm>
+      </AnimateHeight>
+      <h5>{t('patient-card.medical')}</h5>
+      <Wrapper>
+        <Input.Search
+          size="large"
+          placeholder={`${t('patient-card.search-input-placeholder')}`}
+          onChange={(e) => {
+            setTextValue(e.target.value);
+          }}
+          loading={isFetching}
+        />
+        <StyledSelect
+          defaultValue={defaultOrder}
+          options={options}
+          size="large"
+          onChange={(changeText) => {
+            setSearchParams({ order: changeText as string });
+          }}
+        />
+      </Wrapper>
+      <PatientNotes
+        isFetching={isFetching}
+        notes={notes?.data?.notes}
+        total={notes?.data?.total}
+      />
     </>
   );
 }
